+++ koin-projects/examples/coffee-maker/src/main/kotlin/org/koin/example/CoffeeApp.kt
@@ -4,6 +4,7 @@ import org.koin.core.KoinComponent
 import org.koin.core.context.startKoin
 import org.koin.core.context.stopKoin
 import org.koin.core.inject
+import org.koin.core.logger.Level
 import org.koin.core.time.measureDuration
 
 class CoffeeApp : KoinComponent {
@@ -13,7 +14,7 @@ class CoffeeApp : KoinComponent {
 
 fun main() {
     startKoin {
-        printLogger()
+        printLogger(Level.DEBUG)
         modules(listOf(coffeeAppModule))
     }
     val coffeeShop = CoffeeApp()