+++ koin-projects/koin-test-core/src/main/kotlin/org/koin/test/check/CheckModulesDSL.kt
@@ -18,6 +18,7 @@ package org.koin.test.check
 import org.koin.core.Koin
 import org.koin.core.parameter.DefinitionParameters
 import org.koin.core.qualifier.Qualifier
+import org.koin.test.mock.MockProvider
 import kotlin.reflect.KClass
 
 data class CheckedComponent(val qualifier: Qualifier? = null, val type: KClass<*>)
@@ -34,6 +35,7 @@ class ParametersBinding(val koin: Koin) {
             parametersCreators.put(CheckedComponent(qualifier, clazz), creator)
 
     inline fun <reified T : Any> defaultValue(t: T) = defaultValues.put(T::class.java.simpleName, t)
+    inline fun <reified T : Any> defaultValue() = defaultValues.put(T::class.java.simpleName, MockProvider.makeMock<T>())
 }
 
 typealias ParametersCreator = (Qualifier?) -> DefinitionParameters